upd OAMapTopPlace placeId

+++ Sources/Controllers/Map/Layers/OAMapTopPlace.h
@@ -14,12 +14,12 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface OAMapTopPlace : NSObject
 
-@property (nonatomic, readonly) NSInteger placeId;
+@property (nonatomic, readonly) int32_t placeId;
 @property (nonatomic, assign) OsmAnd::PointI position;
 @property (nonatomic, readonly, nullable) UIImage *image;
 @property (nonatomic, readonly) BOOL alreadyExists;
 
-- (instancetype)initWithPlaceId:(NSInteger)placeId
+- (instancetype)initWithPlaceId:(int32_t)placeId
                        position:(const OsmAnd::PointI &)position
                           image:(nullable UIImage *)image
                   alreadyExists:(BOOL)alreadyExists NS_DESIGNATED_INITIALIZER;
